requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
pinecone-client==2.2.4
PyMuPDF==1.23.9
requests==2.31.0
aiohttp==3.9.1
python-multipart==0.0.6
